src/ai/scenario_generator.py
"""
AI-Powered Scenario Generator
Uses Claude API to generate realistic LFCS scenarios
"""

import json
import aiohttp
from typing import Dict, List

class AIScenarioGenerator:
    """Generates scenarios using Claude API"""
    
    def __init__(self):
        self.api_url = "https://api.anthropic.com/v1/messages"
        self.model = "claude-sonnet-4-20250514"
    
    async def generate_scenario(self, category: str, difficulty: str, topic: str) -> Dict:
        """Generate a new scenario using AI"""
        
        prompt = self._build_scenario_prompt(category, difficulty, topic)
        
        payload = {
            "model": self.model,
            "max_tokens": 2000,
            "messages": [
                {
                    "role": "user",
                    "content": prompt
                }
            ]
        }
        
        async with aiohttp.ClientSession() as session:
            async with session.post(
                self.api_url,
                headers={"Content-Type": "application/json"},
                json=payload
            ) as response:
                data = await response.json()
                return self._parse_scenario_response(data)
    
    def _build_scenario_prompt(self, category: str, difficulty: str, topic: str) -> str:
        """Build the prompt for scenario generation"""
        return f"""Generate a realistic LFCS exam scenario for the following:

Category: {category}
Difficulty: {difficulty}
Topic: {topic}

Return ONLY a JSON object with this exact structure (no markdown, no backticks):
{{
    "id": "unique_id",
    "title": "Brief title",
    "description": "Detailed task description (what the user needs to do)",
    "category": "{category}",
    "difficulty": "{difficulty}",
    "topic": "{topic}",
    "points": 10,
    "time_limit": 300,
    "prerequisites": ["list", "of", "setup", "commands"],
    "validation_steps": [
        {{
            "description": "What to check",
            "command": "command to run",
            "expected_output": "expected result",
            "points": 5
        }}
    ],
    "hints": ["hint1", "hint2"],
    "solution_explanation": "How to solve this task"
}}

Make the scenario realistic and practical, similar to actual LFCS exam questions."""
    
    def _parse_scenario_response(self, response: Dict) -> Dict:
        """Parse the AI response and extract scenario"""
        # TODO: Implement response parsing
        pass
