requirements.txt
# Core Dependencies
pyyaml>=6.0
python-dotenv>=1.0.0

# Docker
docker>=6.0.0

# AI/API (optional)
openai>=1.0.0
anthropic>=0.3.0

# Testing
pytest>=7.0.0
hypothesis>=6.0.0
pytest-cov>=4.1.0

# Utilities
tabulate>=0.9.0
colorama>=0.4.6
jinja2>=3.1.0
